+++ src/activemasternode.h
@@ -14,6 +14,7 @@
 #include "net.h"
 #include "sync.h"
 #include "wallet/wallet.h"
+#include "crypto/bls.h"
 
 #define ACTIVE_MASTERNODE_INITIAL 0 // initial state
 #define ACTIVE_MASTERNODE_SYNC_IN_PROCESS 1
@@ -45,6 +46,7 @@ public:
     CPubKey pubKeyMasternode;
 
     std::string strMasterNodePrivKey {""};
+    CBLSSecretKey blsKeyMasternode;
 
     // Initialized while registering Masternode
     Optional<CTxIn> vin;